refactor : 불필요한 style 삭제
