--- src/style/InputContainerStyle.ts
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const InputContainer = styled.div`
-  display: flex;
-  gap: 10px;
-`;
